[AIRFLOW-6001] Lazy load CLI commands (#6594)
* [AIRFLOW-YYY] Lazy load API Client

* [AIRFLOW-YYY] Introduce order in CLI's function names

* [AIRFLOW-YYY] Create cli package

* [AIRLFOW-YYY] Move user and roles command to seperate files

* [AIRLFOW-YYY] Move sync_perm command to seperate file

* [AIRLFOW-YYY] Move task commands to separate file

* [AIRLFOW-YYY] Move pool commands to separate file

* [AIRLFOW-YYY] Move variable commands to separate file

* [AIRLFOW-YYY] Move db commands to separate file

* fixup! [AIRLFOW-YYY] Move variable commands to separate file

* [AIRLFOW-YYY] Move connection commands to separate file

* [AIRLFOW-YYY] Move version command to separate file

* [AIRLFOW-YYY] Move scheduler command to separate file

* [AIRLFOW-YYY] Move worker command to separate file

* [AIRLFOW-YYY] Move webserver command to separate file

* [AIRLFOW-YYY] Move dag commands to separate file

* [AIRLFOW-YYY] Move serve logs command to separate file

* [AIRLFOW-YYY] Move flower command to separate file

* [AIRLFOW-YYY] Move kerberos command to separate file

* [AIRFLOW-YYY] Lazy load CLI commands

* [AIRFLOW-YYY] Fix migration

* fixup! [AIRFLOW-YYY] Fix migration

* fixup! fixup! [AIRFLOW-YYY] Fix migration
