+++ airflow/cli/commands/serve_logs_command.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,23 +14,29 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-API Client that allows interact with Airflow API
-"""
-from importlib import import_module
-from typing import Any
 
-from airflow import api, conf
-from airflow.api.client.api_client import Client
+"""Serve logs command"""
+import os
+
+from airflow import conf
+from airflow.utils import cli as cli_utils
 
 
-def get_current_api_client() -> Client:
-    """
-    Return current API Client depends on current Airflow configuration
-    """
-    api_module = import_module(conf.get('cli', 'api_client'))  # type: Any
-    api_client = api_module.Client(
-        api_base_url=conf.get('cli', 'endpoint_url'),
-        auth=api.API_AUTH.api_auth.CLIENT_AUTH
-    )
-    return api_client
+@cli_utils.action_logging
+def serve_logs(args):
+    """Serves logs generated by Worker"""
+    print("Starting flask")
+    import flask
+    flask_app = flask.Flask(__name__)
+
+    @flask_app.route('/log/<path:filename>')
+    def serve_logs(filename):  # pylint: disable=unused-variable, redefined-outer-name
+        log = os.path.expanduser(conf.get('core', 'BASE_LOG_FOLDER'))
+        return flask.send_from_directory(
+            log,
+            filename,
+            mimetype="application/json",
+            as_attachment=False)
+
+    worker_log_server_port = int(conf.get('celery', 'WORKER_LOG_SERVER_PORT'))
+    flask_app.run(host='0.0.0.0', port=worker_log_server_port)
