+++ tests/cli/commands/test_db_command.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,23 +14,27 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-API Client that allows interact with Airflow API
-"""
-from importlib import import_module
-from typing import Any
 
-from airflow import api, conf
-from airflow.api.client.api_client import Client
+import unittest
+from unittest import mock
+
+from airflow.bin import cli
+from airflow.cli.commands import db_command
 
 
-def get_current_api_client() -> Client:
-    """
-    Return current API Client depends on current Airflow configuration
-    """
-    api_module = import_module(conf.get('cli', 'api_client'))  # type: Any
-    api_client = api_module.Client(
-        api_base_url=conf.get('cli', 'endpoint_url'),
-        auth=api.API_AUTH.api_auth.CLIENT_AUTH
-    )
-    return api_client
+class TestCliDb(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.parser = cli.CLIFactory.get_parser()
+
+    @mock.patch("airflow.cli.commands.db_command.db.initdb")
+    def test_cli_initdb(self, initdb_mock):
+        db_command.initdb(self.parser.parse_args(['db', 'init']))
+
+        initdb_mock.assert_called_once_with()
+
+    @mock.patch("airflow.cli.commands.db_command.db.resetdb")
+    def test_cli_resetdb(self, resetdb_mock):
+        db_command.resetdb(self.parser.parse_args(['db', 'reset', '--yes']))
+
+        resetdb_mock.assert_called_once_with()
